+++ src/scicloj/ml/smile/regression.clj
@@ -250,7 +250,7 @@
   [feature-ds label-ds options]
   (let [entry-metadata (model-type->regression-model
                         (model/options->model-type options))
-        _ (malli/check-schema (:options entry-metadata) options)
+        ;_ (malli/check-schema (:options entry-metadata) options)
         target-colnames (ds/column-names label-ds)
         feature-colnames (ds/column-names feature-ds)
         _ (when-not (= 1 (count target-colnames))
@@ -309,7 +309,9 @@
 (doseq [[reg-kwd reg-def] regression-metadata]
   (let [model-opts {:thaw-fn thaw
                     :hyperparameters (:gridsearch-options reg-def)
-                    :options (:options reg-def)
+                    :options
+                    (malli/options->malli
+                     (:options reg-def))
                     :documentation {:javadoc (class->smile-url (:class reg-def))
                                     :user-guide (-> reg-def :documentation :user-guide)}}
         model-opts (assoc-some model-opts